+++ e2e/jest-resolver-tsconfig-paths/__tests__/module-a.spec.ts
@@ -1,0 +1,7 @@
+import k from '@test/module-a/k';
+import moduleA from '@test/module-a';
+
+it('module-a', () => {
+  expect(k).toBeDefined();
+  expect(moduleA).toBe('module-a');
+});