+++ e2e/jest-resolver-tsconfig-paths/modules/module-a/index.ts
@@ -1,0 +1,3 @@
+const moduleName = 'module-a';
+
+export default moduleName;