+++ e2e/jest-resolver-tsconfig-paths/modules/module-b/index.test.ts
@@ -1,0 +1,5 @@
+import moduleA from '@test/module-a';
+
+it('moduleA', () => {
+  expect(moduleA).toBe('module-a');
+});